renames parser to discogsparser to avoid conflict with python's parser module
