+++ speedup/discogsparser.py
